Former-commit-id: 1fe6827d4af479a7edd1633c6ffce14672229f14 [formerly 06070c0381c3d3931785fbc75fd31c38a3764124]
Former-commit-id: 3bcb0f8fe37cf1f3a6a39f7e860019cd26a3a2a3

+++ app/Http/Controllers/ImportationController.php
@@ -18,6 +18,7 @@ use App\Contact;
 use App\FailRate;
 use App\Currency;
 use App\Contract;
+use PrvValidation;
 use App\Direction;
 use App\Surcharge;
 use PrvSurchargers;
@@ -5301,12 +5302,23 @@ class ImportationController extends Controller
 
     public function DestroyAccount($id){
         try{
-            $account = AccountFcl::find($id);
-            Storage::disk('FclAccount')->delete($account->namefile);
-            $account->delete();
-            return 1;
+            $contract = Contract::where('account_id',$id)->first();
+            if(count($contract) == 1){
+                $data = PrvValidation::ContractWithJob($contract->id);
+                if($data['bool'] == false){
+                    $account = AccountFcl::find($id);
+                    Storage::disk('FclAccount')->delete($account->namefile);
+                    $account->delete();
+                }
+                return response()->json(['success' => 1,'jobAssociate' => $data['bool']]);
+            } else {
+                $account = AccountFcl::find($id);
+                Storage::disk('FclAccount')->delete($account->namefile);
+                $account->delete();
+                return response()->json(['success' => 1,'jobAssociate' => false]);
+            }
         } catch(Exception $e){
-            return 2;
+            return response()->json(['success' => 2,'jobAssociate' => false]);
         }
     }
